requirements.txt
pydantic_ai 
beautifulsoup4 
pandas
python-dotenv